dev: rename dot-key-value-ng to dot-key-value
