+++ core-web/libs/edit-content/src/lib/fields/dot-edit-content-key-value/dot-edit-content-key-value.component.html
@@ -1,4 +1,4 @@
-<dot-key-value-ng
+<dot-key-value
     [variables]="initialValue"
     [autoFocus]="false"
     [showHiddenField]="false"
