+++ core-web/libs/ui/src/lib/components/dot-key-value/dot-key-value-table-row/dot-key-value-table-row.component.html
