+++ core-web/libs/ui/src/lib/components/dot-key-value/dot-key-value.component.scss
