+++ core-web/libs/ui/src/lib/components/dot-key-value/dot-key-value.component.ts
@@ -24,9 +24,9 @@ export interface DotKeyValue {
 }
 
 @Component({
-    selector: 'dot-key-value-ng',
-    styleUrls: ['./dot-key-value-ng.component.scss'],
-    templateUrl: './dot-key-value-ng.component.html',
+    selector: 'dot-key-value',
+    styleUrls: ['./dot-key-value.component.scss'],
+    templateUrl: './dot-key-value.component.html',
     standalone: true,
     imports: [
         CommonModule,
